requirements.txt
django~=4.2.3
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
mysqlclient==1.4.6
requests == 2.7.0 
